% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_app_connection.R
\name{get_app_connection}
\alias{get_app_connection}
\title{Get the app connection from the environment}
\usage{
get_app_connection()
}
\value{
returns a the app connection to the app data.
}
\description{
Call the app connection. The connection is set in the function \code{set_app_connection}.
}
