% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nf_data.R
\name{get_nf_data}
\alias{get_nf_data}
\title{Get notification data from OpenAppBuilder}
\usage{
get_nf_data(
  site = get_app_connection(),
  filter = FALSE,
  filter_variable = NULL,
  filter_variable_value = NULL
)
}
\arguments{
\item{site}{The name of the PostgreSQL database connection (using \code{DBI::dbConnect} or \code{set_app_connection()}).}

\item{filter}{A logical value indicating whether to filter the data (defaults to \code{FALSE}).}

\item{filter_variable}{A character string representing the name of the column to filter if \code{filter == TRUE} and \code{filter_variable_value} is provided.}

\item{filter_variable_value}{A character string representing the value of the \code{filter_variable} column to filter if \code{filter == TRUE}.}
}
\value{
A data frame containing notification interaction data from the PostgreSQL database, with the \code{notification_meta} column parsed into separate columns.
}
\description{
This function retrieves data from the \code{app_notification_interaction} table in OpenAppBuilder and efficiently parses the \code{notification_meta} column from JSON format.
}
\examples{
# TODO

}
