% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nf_data.R
\name{get_nf_data}
\alias{get_nf_data}
\title{Get notification data from Postgres site}
\usage{
get_nf_data(site)
}
\arguments{
\item{site}{Connection to database to get original data (using DBI::dbConnect).}
}
\value{
Notification data from Postgres
}
\description{
Get notification data from Postgres site
}
\examples{
#TODO
}
