% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nf_data.R
\name{get_nf_data}
\alias{get_nf_data}
\title{Retrieve data from the app_notification_interaction table}
\usage{
get_nf_data(
  site = get_app_connection(),
  filter = FALSE,
  UIC_Tracker = NULL,
  app_user_id = "app_user_id",
  country = "Tanzania",
  study
)
}
\arguments{
\item{site}{The name of the PostgreSQL database (using \code{DBI::dbConnect} or \code{set_app_connection()}).}

\item{filter}{A logical value indicating whether to filter the UIC data frame (defaults to \code{FALSE}).}

\item{UIC_Tracker}{A data frame containing UIC data (if filter is \code{TRUE}).}

\item{app_user_id}{A character string representing the name of the column containing app user IDs in \code{UIC_Tracker}.}

\item{country}{A character string representing the country for which to retrieve data (required if filter is \code{TRUE}).}

\item{study}{A character string representing the study for which to retrieve data (required if filter is \code{TRUE}).}
}
\value{
Data frame containing notification data from postgres
}
\description{
This function retrieves data from the `app_notification_interaction`` table in a PostgreSQL database.
}
\examples{
#TODO
}
