% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user_data.R
\name{get_user_data}
\alias{get_user_data}
\title{Get user data from postgres site}
\usage{
get_user_data(
  site,
  date_from,
  date_to = NULL,
  format_date = "\%Y-\%m-\%d",
  tzone_date = "UTC",
  include_UIC_data = TRUE,
  merge_check = TRUE,
  UIC_Tracker,
  app_user_id = "app_user_id",
  join_UIC = "UIC",
  max_dist = 5
)
}
\arguments{
\item{site}{Connection to database to get original data (using DBI::dbConnect).}

\item{date_from}{Date to read in data from.}

\item{date_to}{Date to read in data to.}

\item{format_date}{Character string giving a date-time format as used by \code{\link[base]{strptime}}.}

\item{tzone_date}{Time zone specification to be used for the conversion, if one is required.
"UTC" by default. System-specific (see \code{\link[base]{as.POSIXlt}}), but "" is the current time zone, and "GMT" is UTC (Universal Time, Coordinated).
Invalid values are most commonly treated as UTC, on some platforms with a warning.}

\item{include_UIC_data}{logical. Default \code{TRUE}. Whether to merge UIC data.}

\item{merge_check}{logical. Default \code{TRUE}. Whether to display the close matches when merging UIC data in.}

\item{UIC_Tracker}{If \code{include_UIC_data = TRUE}, UIC tracker data.}

\item{app_user_id}{Default "app_user_id". If \code{include_UIC_data = TRUE}, variable in the postgres data to merge the UIC data by.}

\item{join_UIC}{Default "UIC". If \code{include_UIC_data = TRUE}, variable to merge the UIC data by.}

\item{max_dist}{Default \code{5}. Maximum string distance when merging UIC Tracker data.}
}
\value{
User data from postgres
}
\description{
Get user data from postgres site
}
\examples{
# TODO
}
